% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{datamos}
\alias{datamos}
\alias{datamos-package}
\title{datamos}
\usage{
datamos(id)
}
\arguments{
\item{id}{the integer id of dataset to download}
}
\value{
the desired dataset from data.mos.ru
}
\description{
datamos

Download specific dataset from data.mos.ru
}
\details{
Download specific dataset from data.mos.ru
}
\examples{
df <- datamos(673)
attr(df,"info_df")
attr(df,"info_vars")
}
\author{
Boris Demeshev
}

