% Generated by roxygen2 (4.0.2): do not edit by hand
\name{datamos_list}
\alias{datamos_list}
\title{Download the data.frame of all datasets available at data.mos.ru}
\usage{
datamos_list()
}
\value{
the data.frame of all datasets available at data.mos.ru
}
\description{
Download the data.frame of all datasets available at data.mos.ru
}
\details{
Download the data.frame of all datasets available at data.mos.ru
}
\examples{
dsets <- datamos_list()
}

